requirements.txt
psycopg2
fastapi
pydantic
openai==1.55.3 
httpx==0.27.2
python-dotenv==1.0.1